+++ src/converters/piral-million/src/extension.ts
@@ -1,12 +1,4 @@
 import type { ExtensionSlotProps } from 'piral-core';
-import { createElement, useContext } from 'million/react';
-import { piralContext } from './mount';
+import { m } from 'million';
 
-export function createExtension(rootName: string) {
-  return (props: ExtensionSlotProps) => {
-    const { piral } = useContext(piralContext);
-    const m = createElement as any;
-    //TODO
-    return m(rootName);
-  };
-}
+export const Extension = (props: ExtensionSlotProps) => m('piral-extension', props);