+++ src/samples/sample-cross-fx/src/index.tsx
@@ -3,6 +3,7 @@ import 'core-js/es/reflect';
 import 'core-js/stable/reflect';
 import 'core-js/features/reflect';
 import 'zone.js/dist/zone.js';
+import '@angular/compiler';
 import '@webcomponents/webcomponentsjs/webcomponents-loader';
 import '@webcomponents/webcomponentsjs/webcomponents-bundle.js';
 import '@webcomponents/webcomponentsjs/custom-elements-es5-adapter';